+++ src/main/kotlin/lotto/domain/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 interface NumberGenerator {
     fun generate(): List<Int>
