+++ src/main/kotlin/lotto/domain/WinningNumbers.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class WinningNumbers(
     val numbers: List<Int>
